# currency-get-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A web service that, upon request, collects data from the public API of the national bank and saves the data to the local TEST database",
        "title": "Swagger kursRates API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "paths": {
        "/currency/{date}": {
            "get": {
                "description": "Get currency data for a specific date.\nGet currency data for a specific date and currency code.",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "currency",
                    "currency"
                ],
                "summary": "Get currency data by date and code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in DD.MM.YYYY format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/currency/{date}/{code}": {
            "get": {
                "description": "Get currency data for a specific date.\nGet currency data for a specific date and currency code.",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "currency",
                    "currency"
                ],
                "summary": "Get currency data by date and code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in DD.MM.YYYY format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Currency code (e.g., USD)",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}